+++ CHANGES.md
@@ -9,6 +9,7 @@
 - GEMET ingurumen thesaurus-a gehitu da. (#237)
 - Gizarte Zerbitzuen hiztegia gehitu da. (#227)
 - Gotzon Garateren atsotitzak gehitu dira. (#235)
+- Kamusi project hiztegia gehitu da. (#244)
 - Laneki hiztegia konpondu da. (#230)
 - Nafarroako euskalkien hiztegia gehitu da. (#233)
 - Trengintza hiztegia gehitu da. (#228)